requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.5
google_auth_oauthlib==1.0.0
protobuf==3.20.3
PyJWT==2.7.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
Requests==2.31.0
Werkzeug==2.3.6
